generic architecture for both 3d and 2d inputs
